add control to check if img already downloaded

+++ moviespred/data.py
@@ -122,11 +122,12 @@ def movies_pred(genre):
     for content, save_name in zip(list_address, list_names):
         res = rq.get(content)
         save_path = f"{paths['raw_images']}/{genre}/{save_name}.jpg"
-        with open(save_path, 'w') as f:
-            image = Image.open(io.BytesIO(res.content))
-            image.save(f)
-            print(f'{c}/{df_len} downloaded image of {genre}')
-            c = c + 1
+        if not os.path.isfile(save_path):
+            with open(save_path, 'w') as f:
+                image = Image.open(io.BytesIO(res.content))
+                image.save(f)
+                print(f'{c}/{df_len} downloaded image of {genre}')
+                c = c + 1
 
 def remove(genre,dir_):
     directory = os.path.join(dir_, genre)
